+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## v1.1.0
 
 - Adds api for `createBlockDecorator`
+
+## v1.2.0
+
+- Adds api for `composer.toString`
+- Fixes exception when props are passed to a composer with functional modifiers.